requirements.txt
pymysql
Flask-Cors==3.0.7
flask_migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2